requirements.txt
Flask
firebase-admin
google-auth
requests
gunicorn
schematics
